Unify the two role-privilege mappers behind ProviderConfigBase
OidcRolePrivilegeMapper and SamlRolePrivilegeMapper folded the same
ProviderConfigBase members with ~90 lines each and each declared a same-named
nested RoleGrants (defeating IDE navigation), forcing a contributor to diff the
two classes by eye to find the three documented behavioral quirks.

Collapse to one RolePrivilegeMapper.Evaluate(IEnumerable<string>,
ProviderConfigBase) returning one RoleGrants (the SAML caller ignores Valid,
whose decision belongs to SamlLoginPolicy). Every member read already lives on
ProviderConfigBase, so no new interface is needed; both authorize-state builders
call the one mapper.

Reconcile the three quirks to one robust behavior (a login-path authz change,
all toward graceful fail-closed / trusted-config normalization, none reachable
as an over-grant via IdP-controlled input):
- role comparison is null-safe both ways (string.Equals) — a null role or a null
  configured entry is a non-match, not a NullReferenceException;
- the folder-role mapping is null-guarded in both protocols — a null mapping with
  folder roles enabled grants no folders instead of failing the login with a 500;
- the trusted, admin-authored configured map-role is trimmed before comparison in
  both protocols; the IdP-supplied role is compared raw, so there is no
  whitespace-injection vector.

Deletes both mappers and the duplicate RoleGrants, consolidates the
characterization tests into one file reflecting the reconciled behavior, updates
the FsCheck monotonicity properties, and retires the stale #89 mapper comment.

Closes #367
